tab based log timer
