+++ src/services/DateFilters.js
@@ -1,4 +1,4 @@
-import { format } from 'date-fns'
+import { format, distanceInWords } from 'date-fns'
 
 export function toLocalISO (value) {
   return format(value)
@@ -7,3 +7,5 @@ export function toLocalISO (value) {
 export function toFullDateTime (value) {
   return format(value, 'dddd, MMMM Do, YYYY h:mm:ss A')
 }
+
+export const differenceInWords = (value, compareDate) => distanceInWords(value, compareDate)
